requirements.txt
google-genai>=1.50.0
markdown>=3.6
